requirements.txt
streamlit==1.32.0
twelvelabs>=0.4.0
python-dotenv==1.0.1
pandas==2.2.0
plotly==5.19.0